upload Mac package under work
This package is almost identical to src_1.0 but it has a few changes
mostly related to the GUI. Mac is having many compatibility issues with
wxpython. It works but it keeps throwing errors sometimes and it needs
to be fully debugged. Once that is done. The same code should be run on
Windows and Linux to test their compatibility